functions/list-wallets/requirements.txt
ydb==3.8.3
PyJWT==2.8.0